+++ app/src/main/AndroidManifest.xml
@@ -18,10 +18,10 @@
             android:label="@string/title_activity_home"
             android:theme="@style/Theme.HyperAppDev" />
         <activity
-            android:name=".LibAuthentication.AuthActivities.Authentication"
+            android:name="com.example.carcab.LibAuthentication.AuthActivities.Authentication"
             android:exported="false" />
         <activity
-            android:name=".LibAuthentication.AuthActivities.MainActivity"
+            android:name="com.example.carcab.LibAuthentication.AuthActivities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
