+++ app/src/main/java/com/example/carcab/LibAuthentication/AuthActivities/Authentication.java
@@ -1,6 +1,5 @@
-package com.example.hyperapp_dev.LibAuthentication.AuthActivities;
+package com.example.carcab.LibAuthentication.AuthActivities;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.activity.EdgeToEdge;
@@ -9,30 +8,26 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-import com.example.hyperapp_dev.R;
+import com.example.carcab.R;
 
-public class MainActivity extends AppCompatActivity {
+public class Authentication extends AppCompatActivity {
+
+    /*
+    TODO: Reminder
+    This activity must have two fragments, Login and Register fragments.
+    Firebase authentication MAY be used here, or SharedPreferences (maybe), the former being
+    to comply with using Firebase as part of the task requirement
+     */
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_authentication);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
-
-        toAuthenticationScreen();
-    }
-
-    // TODO: Remove and modify whatever is necessary (e.g. modify/delete the activity
-    // TODO: Whatever the changes is, be sure to check over the manifest file
-    // Do modify/omit this when the application has a flow and navigation
-    private void toAuthenticationScreen()
-    {
-        Intent intent = new Intent(MainActivity.this, Authentication.class);
-        startActivity(intent);
     }
 }
