+++ app/src/main/java/com/example/carcab/LibAuthentication/AuthFragments/UserLogin.java
@@ -1,4 +1,4 @@
-package com.example.hyperapp_dev.LibAuthentication.AuthFragments;
+package com.example.carcab.LibAuthentication.AuthFragments;
 
 import android.os.Bundle;
 
@@ -8,14 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.hyperapp_dev.R;
+import com.example.carcab.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link UserRegister#newInstance} factory method to
+ * Use the {@link UserLogin#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class UserRegister extends Fragment {
+public class UserLogin extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class UserRegister extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public UserRegister() {
+    public UserLogin() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class UserRegister extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment UserRegister.
+     * @return A new instance of fragment UserLogin.
      */
     // TODO: Rename and change types and number of parameters
-    public static UserRegister newInstance(String param1, String param2) {
-        UserRegister fragment = new UserRegister();
+    public static UserLogin newInstance(String param1, String param2) {
+        UserLogin fragment = new UserLogin();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class UserRegister extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user_register, container, false);
+        return inflater.inflate(R.layout.fragment_user_login, container, false);
     }
 }
